Migrate characters to app router
