+++ components/PaginationButtons.js
@@ -1,8 +1,29 @@
+'use client';
+
 import Button from './Button';
+import { useRouter } from 'next/navigation';
+import { generateQuery } from '../utils';
 import styles from '../styles/PaginationButtons.module.css';
 
 const PaginationButtons = (props) => {
-  const { changePage, currentPage } = props;
+  const { info, currentPage } = props;
+
+  const router = useRouter();
+
+  const changePage = (type) => {
+    if (!info[type]) {
+      return;
+    }
+
+    const searchParams = new URL(info[type]).searchParams;
+    const page = searchParams.get('page');
+
+    let url = `/characters?`;
+
+    url += generateQuery([{ label: 'page', value: page }]);
+
+    router.push(url);
+  };
 
   return (
     <div className={styles.wrapper}>
